Changed the int to size_t in linkedList to avoid the indices being lower than 0

+++ include/linkedList.h
@@ -18,7 +18,7 @@ struct LinkedList
 {
     Node *start;
     Node *end;
-    int listLength;
+    size_t listLength;
 };
 
 void LinkedList_Print(LinkedList*, void (*printFunc)(void*));
